Initial reorganize and modification
Move as much of the non-reference content to the bottom of the file.
Should be easier to access the style guide without having to read about
contributors, etc.

Reduced the number of 'bad' examples to reduce the noise. Focus on
examples that conform to the style guide.

Modify the examples as needed to match my personal preferences, e.g.,
multiple var statements, single quotes, and so on.

Still about half the document to look over.